move entity into a syntax
